fixing up tests
